trynow: Show login button instead of launch button while logged out
- Refactor terms and conditions into a separate component
- Add login button component
- Simplify launch app component
- Simplify calculation of the need to show login button and information

Contributed on behalf of STMicroelectronics
