+++ node/try-now-page/src/components/LoginButton.tsx
@@ -1,0 +1,9 @@
+export interface LoginButtonProps {
+  login: () => void;
+}
+export const LoginButton: React.FC<LoginButtonProps> = ({login}: LoginButtonProps) => (<button
+  className='App__try-now-button'
+  onClick={() => login()}
+>
+    Login
+</button>);
